requirements.txt
Deprecated==1.2.18
img2pdf==0.6.1
lxml==6.0.0
packaging==25.0
pdf2image==1.17.0
pikepdf==9.10.0
pillow==11.3.0
wrapt==1.17.2
